Move rewriter logic to a module and set hook for prepareStackTrace
